refactor: share seccomp wrapper config builder
